refactor: split linker category tasks
